// angularProject/src/app/api/model/registerModelDto.ts
/**
 * TrueValue
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegisterModelDto { 
    email?: string | null;
    username?: string | null;
    password?: string | null;
    gender?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    dateOfBirth?: string | null;
    phoneNumber?: string | null;
    is18OrAbove?: boolean;
    isWhatsAppNumber?: boolean;
    countryId?: string | null;
}